Minimum spanning tree (algorythm)
Minimum spanning tree. That is a program which takes a connected, undirected graph with weights and outputs the minimum spanning tree of the graph i.e., a subgraph that is a tree, contains all the vertices, and the sum of its weights is the least possible. How it works:To find the minimum spanning tree (MST) of a connected, undirected graph with weights, we can use Prim's or Kruskal's algorithm. Below is a Python program that implements Kruskal's algorithm to find the MST. This program uses the networkx library, which provides convenient functions for graph manipulation. Important: make sure you have the networkx library installed. You can install it using pip if you haven't done so: pip install networkx Explanation of the Code: Imports: The program imports the networkx library, which is used for creating and manipulating graphs. Creating the Graph: The create_graph function takes a list of edges (each defined by two nodes and a weight) and creates a weighted graph. Finding the MST: The minimum_spanning_tree function uses the networkx function minimum_spanning_tree to compute the MST of the graph. Main Function: In the main function, we define a list of edges for the graph, create the graph, compute the MST, and print the edges of the MST along with their weights.