Site Checker with Time Scheduling
An application that attempts to connect to a website or server every so many minutes or a given time and check if it is up. If it is down, it will notify you by email or by posting a notice on screen.
This program will attempt to connect to a specified website at regular intervals and notify you if the site is down. You can modify the email notification part to suit your needs. To run this program, you'll need the requests library for checking the website status and smtplib for sending email notifications. You can install the requests library using pip if you haven't done so: pip install requests (put that in the command line) 
Brief explanation of the code: 
SiteChecker Class: This class is responsible for checking the website status, notifying the user, and sending email notifications. check_site: Attempts to connect to the specified URL and checks the response status. notify_user: Prints a message to the console and sends an email notification if the site is down. send_email_notification: Sends an email using SMTP when the site is down. run: Continuously checks the site at the specified interval. Main Function: Initializes the SiteChecker with the URL to check, the interval for checking, and the email information. It then starts the site checking process. Running the Program: Replace the placeholder values in the email_info dictionary with your actual email credentials. Replace the url variable with the website you want to monitor. Run the program, and it will check the site at the specified interval, notifying you if it goes down. 
Note: Make sure to enable "Less secure app access" in your Google account settings if you're using Gmail for sending emails.