Gif Creator
Creating a GIF  in Python is  accomplished by using the Pillow library, which is a good tool for image manipulation, along with the image io library for saving images as GIFs. 
Make sure you have Pillow and imageio installed. You can install them using pip: pip install Pillow imageio
The program will:
Prompt the user to input the file paths of the images they wish to include in the GIF.
Allow the user to specify frame duration (i.e. how long each frame should appear in the GIF).
Create a GIF and save it to the specified output path.
File Size Check: Before adding each image to the list, check if its size is less than 35 KB. If an image is larger, prompt the user to enter a new file path until a valid file is provided.
Loop Count Input: Ask the user for the number of times the GIF should loop.